requirements.txt
albumentations==1.1.0
fire==0.4.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
numpy==1.21.2
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
pandas==1.3.4
Pillow==9.0.1
pytorch-lightning==1.6.0
PyYAML==6.0
scikit-image==0.19.2
scikit-learn==1.1.1
scikit-learn-intelex==2021.6.3
scipy==1.6.0
seaborn==0.11.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-metadata==1.7.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2022.3.25
timm==0.5.4
tokenizers==0.12.1
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.4.1
torchvision==0.12.0
tornado==6.1
transformers==4.18.0